ZapVisualSimDyn : improve the project compilation
The project main directory is cleaned. The README file
is updated to explain to the user how to compile
the project easily.
A typedef in the console.h file is modified to allow
the project to compile on 64bits systems.
There is many segmentation faults when the project is
executed...